refactor(tests): split RunBotCoreTests.swift catch-all by domain (#2990)
Phase 3 intermediate-granularity split; suites moved verbatim (names,
comments, issue refs, and test bodies unchanged):

- Runner/Models/RunnerModelContractTests.swift
  (ModelElapsedIntegration, ActiveJob.isLocalRunner,
   RunnerModel presentation, JobStatus.isActive, JobConclusion.isFailure)
- Runner/Polling/PollResultBuilderTests.swift
  (PollResultBuilder, buildGroupState #1041, evictFreshShas #2444)
- Platform/ProcessRunnerTests.swift (ProcessRunner.runAsync stdin)

TestModelHelpers.swift moves to existing TestSupport/ — it is consumed by
multiple destination files. No SIGTERM test reinstated. Catch-all file
deleted. Test count unchanged: 118 tests in 27 suites.
